move ld-ldf-reader into tools

+++ tools/ld-ldf-reader/ld-ldf-reader.c
@@ -88,7 +88,7 @@ static int decode_packet(AVCodecContext *dec, const AVPacket *pkt)
         // Write the raw audio data samples to stdout
         size_t rv = fwrite(frame->extended_data[0] + offset, sizeof(uint8_t), length, stdout);
         if (rv != length) {
-            fprintf(stderr, "write error %ld", offset);
+            fprintf(stderr, "write error %lld", (long long)offset);
             return -1;
         }
 
@@ -204,7 +204,7 @@ int main (int argc, char **argv)
 
     fprintf(stderr, "RATE:%d\n", audio_dec_ctx->sample_rate);
     // From fmt_ctx, this is the approximate length in ms.  (divide by 1000 for actual time)
-    fprintf(stderr, "DURATION:%ld\n", fmt_ctx->duration);
+    fprintf(stderr, "DURATION:%lld\n", (long long)fmt_ctx->duration);
 
     frame = av_frame_alloc();
     if (!frame) {
